Project Setup
This program will simulate a banking system that will pull information from an external CSV file and user will be able to make changes to it. The user will be allowed to access account information saved in the .csv and modify it as desired. A log file will be created after every program execution. Classes file and main file created as skeleton for the rest of the code.